pyproject.toml
[project]
name = "tft-m5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "pytorch-lightning>=2.0.0",
    "PyYAML>=6.0",
    "matplotlib>=3.10.7",
]
